[tests] Update tests to make native type vs normal type comparison explicit (#9097)
We've modified our NUnitLite to special-case the native types, so that they
can easily be compared with the standard types, but that doesn't work anymore
when switching to the official NUnit[Lite], so change all asserts to
explicitly convert whenever needed.

+++ tests/apitest/src/AppKit/NSSlider.cs
@@ -23,7 +23,7 @@ namespace Xamarin.Mac.Tests
 #else
 			slider.IsVertical = 1;
 #endif
-			Assert.AreEqual (1, slider.IsVertical);
+			Assert.AreEqual ((nint) 1, slider.IsVertical);
 		}
 	}
 }
