+++ tests/monotouch-test/CoreML/MLMultiArrayTest.cs
@@ -71,7 +71,7 @@ namespace MonoTouchFixtures.CoreML
 			var shape = new nint [] { 10 };
 			using (var arr = new MLMultiArray (shape, MLMultiArrayDataType.Int32, out err)) {
 				Assert.IsNull (err, "err");
-				Assert.AreEqual (10, arr.Count, "Count");
+				Assert.AreEqual ((nint) 10, arr.Count, "Count");
 				Assert.AreEqual (new nint [] { 10 }, arr.Shape, "Shape");
 				Assert.AreEqual (new nint [] { 1 }, arr.Strides, "Strides");
